Update Main.cpp
Changed ITEM to GET as the verb for picking up an item

Commented in some pseudo code that can be used to help build the commands needed to get the key and lock functionality working when it is coded it. Also added some dialog in the comments as well. This is under the USE verb input validation.